export search bar
